requirements.txt
Flask==2.2.2
gunicorn==20.1.0
requests==2.28.2
beautifulsoup4==4.11.1
selenium==4.10.0
webdriver-manager==3.8.6
# Optional: Uncomment if you install Whisper for transcription
openai-whisper==20230314
